+++ experiments/default/experiment.sh
@@ -7,13 +7,20 @@ TORTE_REVISION=main; [[ $TOOL != torte ]] && builtin source /dev/stdin <<<"$(cur
 # This experiment extracts, transforms, and solves a single feature model.
 # It serves as a demo and integration test for torte and also returns some common statistics of the model.
 
+# here we define global configuration variables, which are explicitly passed to stages below
 TIMEOUT=10
 JOBS=4
 
-add-payload-file evaluation.ipynb
+add-payload-file evaluation.ipynb # we can add Jupyter notebooks to execute at the end of the experiment
+add-payload-file Kconfig.test # we can inject individual KConfig files to test extraction on small examples
+download-payload-file smart_home_fm.uvl https://www.uvlhub.io/hubfiles/download/189 # we can also download arbitrary payload files from the web
+download-payload-file Tankwar.dimacs \
+    https://raw.githubusercontent.com/SoftVarE-Group/feature-model-benchmark/refs/heads/master/feature_models/dimacs/games/Tankwar/Schulze2012.dimacs
 
 experiment-systems() {
-    add-busybox-kconfig-history --from 1_36_0 --to 1_36_1
+    add-busybox-kconfig-history --from 1_36_0 --to 1_36_1 # usually, we add (excerpts of) system histories to analyze here
+    # add-payload-file-kconfig Kconfig.test # we can also just analyze one KConfig file, which will be parsed with the original LKC implementation in Linux
+    # (the previous line is commented out here for demonstration purposes, because it will cause Linux to be cloned, which takes some time and space)
 }
 
 experiment-test-systems() {
@@ -23,7 +30,16 @@ experiment-test-systems() {
 experiment-stages() {
     clone-systems
     read-statistics
+
+    # usually we extract KConfig models from ground truth ...
     extract-kconfig-models
+
+    # ... but we can also inject pre-existing feature model files directly
+    # here, any files are legal that can be successfully parsed by subsequent stages (typically FeatJAR or FeatureIDE)
+    # this allows us to integrate with feature-model repositories, such as the feature-model benchmark or UVLHub (where we download these files from)
+    # inject-kconfig-model --payload-file smart_home_fm.uvl
+    # inject-kconfig-model --payload-file Tankwar.dimacs
+    # (the previous lines are commented out here for demonstration purposes, to focus on the BusyBox model)
 
     transform-model-with-featjar --transformer transform-model-to-xml-with-featureide --output-extension xml --timeout "$TIMEOUT"
     transform-model-with-featjar --transformer transform-model-to-uvl-with-featureide --output-extension uvl --timeout "$TIMEOUT"